qqbot/requirements.txt
requests
python-dotenv
websockets